feat: 이름으로 회원 조회 기능 구현
