+++ src/main/java/com/project/hireup/jwt/JwtTokenProvider.java
@@ -36,7 +36,7 @@ public class JwtTokenProvider {
   // 암호화 알고리즘
   private final SignatureAlgorithm signatureAlgorithm = SignatureAlgorithm.HS256;
 
-  private final long tokenTime = 1000L * 60 * 60; // 1시간
+  private final long tokenTime = 1000L * 60 * 60 * 24; // 1시간
 
   // Bean 생성 후 자동 실행 (secretKey를 Key 객체로 변환하여 저장)
   @PostConstruct
