+++ src/main/java/com/project/hireup/service/AuthService.java
@@ -1,0 +1,137 @@
+package com.project.hireup.service;
+
+import static com.project.hireup.type.ErrorCode.ALREADY_AUTH;
+import static com.project.hireup.type.ErrorCode.EMAIL_UNVERIFIED;
+import static com.project.hireup.type.ErrorCode.INVALID_PASSWORD;
+import static com.project.hireup.type.ErrorCode.NOT_EQUAL_CONFIRM_PASSWORD;
+import static com.project.hireup.type.ErrorCode.NOT_EQUAL_TOKEN;
+import static com.project.hireup.type.ErrorCode.NOT_EXIST_EMAIL;
+import static com.project.hireup.type.ErrorCode.NOT_EXIST_EMAIL_AUTH_KEY;
+import static com.project.hireup.type.ErrorCode.SUSPENDED_USER;
+import static com.project.hireup.type.ErrorCode.USER_ALREADY_EXISTS;
+
+import com.project.hireup.component.MailComponent;
+import com.project.hireup.dto.SignUpRequestDto;
+import com.project.hireup.entity.User;
+import com.project.hireup.exception.HireUpException;
+import com.project.hireup.jwt.JwtTokenProvider;
+import com.project.hireup.repository.UserRepository;
+import com.project.hireup.type.UserRole;
+import com.project.hireup.type.UserStatus;
+import java.util.UUID;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class AuthService {
+
+  private final UserRepository userRepository;
+  private final MailComponent mailComponent;
+  private final PasswordEncoder passwordEncoder;
+  private final JwtTokenProvider jwtTokenProvider;
+
+  @Value("${admin.token}")
+  private String adminToken;
+
+  @Value("${base.url}")
+  private String baseUrl;
+
+  // 회원가입
+  public void signUp(SignUpRequestDto requestDto) {
+
+    // 회원가입 진행 -> 이미 등록된 email이 있다면 예외 처리
+    if (userRepository.existsByEmail(requestDto.getEmail())) {
+      throw new HireUpException(USER_ALREADY_EXISTS);
+    }
+
+    // 비밀번호와 비밀번호 확인 입력이 다를시 예외 처리
+    if (!requestDto.getPassword().equals(requestDto.getConfirmPassword())) {
+      throw new HireUpException(NOT_EQUAL_CONFIRM_PASSWORD);
+    }
+
+    // isAdmin 값이 true 이고 함께 넣어준 토큰 값이 일치한다면 Partner role 부여
+    UserRole role = determineUserRole(requestDto);
+
+    // email 인증 키를 UUID를 사용하여 랜덤 값으로 생성
+    String uuid = UUID.randomUUID().toString();
+
+    User user = User.builder()
+        .email(requestDto.getEmail())
+        .name(requestDto.getName())
+        .password(passwordEncoder.encode(requestDto.getPassword())) // 비밀번호 암호화 적용
+        .emailAuthKey(uuid)
+        .userRole(role)
+        .emailAuthYn(false)
+        .status(UserStatus.UNVERIFIED)
+        .build();
+    userRepository.save(user);
+
+    String email = requestDto.getEmail();
+    String subject = "HireUp 사이트 가입을 환영합니다!";
+    String text = "<p>" + requestDto.getName() + "님, HireUp 사이트 가입을 환영합니다!</p>"
+        + "<p>아래 링크를 클릭하셔서 가입을 완료하세요.</p>"
+        + "<div><a target='_blank' href='" + baseUrl + "/api/auth/email-auth?uuid=" + uuid
+        + "'>가입 완료</a></div>"
+        + "<p>가입을 완료하시면 HireUp의 다양한 서비스를 이용하실 수 있습니다.</p>"
+        + "<p>감사합니다.</p>";
+
+    mailComponent.sendMail(email, subject, text);
+  }
+
+  // 로그인
+  public String signIn(String email, String password) {
+    // 이메일로 사용자 조회
+    User user = userRepository.findByEmail(email)
+        .orElseThrow(() -> new HireUpException(NOT_EXIST_EMAIL));
+
+    // 비밀번호 검증
+    if (!passwordEncoder.matches(password, user.getPassword())) {
+      throw new HireUpException(INVALID_PASSWORD);
+    }
+
+    // 계정 상태 확인
+    if (user.getStatus() == UserStatus.UNVERIFIED) { // 이메일 인증이 되지 않은 유저 로그인 방지
+      throw new HireUpException(EMAIL_UNVERIFIED);
+    }
+    if (user.getStatus() == UserStatus.SUSPENDED) { // 정지된 유저 로그인 방지
+      throw new HireUpException(SUSPENDED_USER);
+    }
+
+    // JWT 토큰 생성
+    return jwtTokenProvider.createToken(user.getEmail(), user.getUserRole().name(),
+        user.getStatus().name());
+  }
+
+  // 이메일 인증
+  public void emailAuth(String uuid) {
+
+    User user = userRepository.findByEmailAuthKey(uuid)
+        .orElseThrow(() -> new HireUpException(NOT_EXIST_EMAIL_AUTH_KEY));
+
+    if (user.isEmailAuthYn()) {
+      throw new HireUpException(ALREADY_AUTH);
+    }
+
+    user.setEmailAuthYn(true);
+    user.setStatus(UserStatus.ACTIVE);
+    userRepository.save(user);
+  }
+
+  // 유저에게 권한 부여하는 메서드 생성
+  private UserRole determineUserRole(SignUpRequestDto requestDto) {
+    if (requestDto.isAdmin()) {
+      validateAdminToken(requestDto.getAdminToken());
+      return UserRole.ROLE_ADMIN;
+    }
+    return UserRole.ROLE_USER;
+  }
+
+  private void validateAdminToken(String adminToken) {
+    if (!this.adminToken.equals(adminToken)) {
+      throw new HireUpException(NOT_EQUAL_TOKEN);
+    }
+  }
+}
